% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{abundence_meter}
\alias{abundence_meter}
\title{Low-level function to calculate abundances for each cluster within each sample}
\usage{
abundence_meter(signalling_type, list_cell_ctDist)
}
\arguments{
\item{list_cell_ctDist}{}
}
\value{

}
\description{
Low-level function to calculate abundances for each cluster within each sample
}
