% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDB.R
\name{add_marker_GDB}
\alias{add_marker_GDB}
\title{Add markers as nodes to neo4j database}
\usage{
add_marker_GDB(use_markers, gdb)
}
\arguments{
\item{gdb}{}
}
\value{

}
\description{
Add markers as nodes to neo4j database
}
