% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDB.R
\name{add_observation_GDB}
\alias{add_observation_GDB}
\title{Allocate observations as nodes based on marker and cell population nodes}
\usage{
add_observation_GDB(gdb)
}
\arguments{
\item{gdb}{}
}
\value{

}
\description{
Allocate observations as nodes based on marker and cell population nodes
}
