% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDB.R
\name{add_phenounite_GDB}
\alias{add_phenounite_GDB}
\title{Allocate phenounite, phenotype units, nodes based on markers and clusters in database}
\usage{
add_phenounite_GDB(gdb)
}
\arguments{
\item{gdb}{}
}
\value{

}
\description{
Allocate phenounite, phenotype units, nodes based on markers and clusters in database
}
