% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDB.R
\name{add_populatio_GDB}
\alias{add_populatio_GDB}
\title{Allocate cell populations within each sample based on clusterisation}
\usage{
add_populatio_GDB(cell_clustering_list, gdb)
}
\arguments{
\item{gdb}{}
}
\value{

}
\description{
Allocate cell populations within each sample based on clusterisation
}
