% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDB.R
\name{add_sample_GDB}
\alias{add_sample_GDB}
\title{Add samples as nodes to neo4j database}
\usage{
add_sample_GDB(fcs_raw, gdb)
}
\arguments{
\item{gdb}{}
}
\value{

}
\description{
Add samples as nodes to neo4j database
}
