% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDB.R
\name{add_signals_between_clusters_GDB}
\alias{add_signals_between_clusters_GDB}
\title{Add correlation between clusters as relationships to the database}
\usage{
add_signals_between_clusters_GDB(signals_between_clusters, gdb)
}
\arguments{
\item{gdb}{}
}
\value{

}
\description{
Add correlation between clusters as relationships to the database
}
