% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDB.R
\name{add_signals_in_cluster_GDB}
\alias{add_signals_in_cluster_GDB}
\title{Add correlation within clusters as relationships to the database}
\usage{
add_signals_in_cluster_GDB(signals_in_cluster, gdb)
}
\arguments{
\item{gdb}{}
}
\value{

}
\description{
Add correlation within clusters as relationships to the database
}
