% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{asinh_transformation}
\alias{asinh_transformation}
\title{Transformation from "count" to "asinh" data}
\usage{
asinh_transformation(fcs_raw, cofactor = 5, use_markers = NULL)
}
\arguments{
\item{fcs_raw}{flowSet object}

\item{cofactor}{digit with used as denominator to transform data}
}
\value{
flowSet object with transform data by asinh function and divided
by cofactor
}
\description{
Transformation from "count" to "asinh" data
}
