% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{backgraund_correlation}
\alias{backgraund_correlation}
\title{Get bg_ctCor_data object with background correlation statistics}
\usage{
backgraund_correlation(list_cell_ctDist, list_expData, method = "spearman")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Get bg_ctCor_data object with background correlation statistics
}
