% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{background_anova}
\alias{background_anova}
\title{Get bg_anova object with background ANOVA statistics}
\usage{
background_anova(list_cell_ctDist, list_expData)
}
\arguments{
\item{list_expData}{}
}
\value{

}
\description{
Get bg_anova object with background ANOVA statistics
}
