% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{cell_filter}
\alias{cell_filter}
\title{Low-level function for comparison of detected correlations between clusters with background correlations}
\usage{
cell_filter(
  gene_ctCor_data,
  bg_ctCor_data,
  bg_anova,
  bg_anova_alpha = 0.01,
  bg_ctCor_alpha = 0.01,
  gene_ctCor_alpha = 0.01
)
}
\arguments{
\item{gene_ctCor_alpha}{}
}
\value{

}
\description{
Low-level function for comparison of detected correlations between clusters with background correlations
}
