% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{cell_type_organiser}
\alias{cell_type_organiser}
\title{Low-level function to get cell_ctDist object for one sample}
\usage{
cell_type_organiser(
  coldata_sce,
  cell_id_column = "rownames",
  cell_type_column = "cluster"
)
}
\arguments{
\item{cell_type_column}{}
}
\value{

}
\description{
Low-level function to get cell_ctDist object for one sample
}
