% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_markers.R
\name{comparator_mk}
\alias{comparator_mk}
\title{Compare background correlation landscape versus target correlations
and extract information whether signals pass-through or get the complete statistic information}
\usage{
comparator_mk(
  gene_to_gene_cor,
  anova_mk,
  bg_corr_mk,
  contrast_corr_land_mk,
  anova_mk_alpha = 0.01,
  bg_corr_mk_alpha = 0.01,
  contrast_corr_land_mk_alpha = 0.01,
  use_markers = NULL,
  statistic_output = F
)
}
\arguments{
\item{statistic_output}{if TRUE, function return complete statistic information. Defaul default is FALSE, return pass or not inf}
}
\value{

}
\description{
Compare background correlation landscape versus target correlations
and extract information whether signals pass-through or get the complete statistic information
}
