% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{ct_sample_aggregator}
\alias{ct_sample_aggregator}
\title{Get list_cell_ctDist object for correlation analysis}
\usage{
ct_sample_aggregator(
  list_coldata_sce,
  cell_id_column = "rownames",
  cell_type_column = "cluster",
  do_cluster_unification = TRUE
)
}
\arguments{
\item{do_cluster_unification}{}
}
\value{

}
\description{
Get list_cell_ctDist object for correlation analysis
}
