% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{cytofBrowser_server}
\alias{cytofBrowser_server}
\title{Creation of server part of Shiny App}
\usage{
cytofBrowser_server(input, output)
}
\arguments{
\item{input}{shiny input object with from shiny UI to server}

\item{output}{shiny output object from server to UI and back to server}
}
\value{

}
\description{
the function forms the body of backend part of Shiny App.
It also shapes the architecture of the pipeline. there are pipeline
blocks here which reflect the steps of CyTOF analysis. All computational
functions are placed in other files. The blocks of this function request
computational functions from that files by suitable order.
}
