% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{cytofCoreGUI}
\alias{cytofCoreGUI}
\title{The main function to run cutofCore graphic interface}
\usage{
cytofCoreGUI()
}
\value{
The function runs the Shiny App in browser.
}
\description{
Run graphical user interface for cytofanalyzer.
The function runs the Shiny App in browser. The current package
was created with assumption that it will be run with GUI mode
as Shiny App, That is why the function is the most appropriate
way to use the cytofanalyzer.
}
\examples{
\dontrun{
cytofCoreGUI()
}
}
