% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusterisation.R
\name{filter_edges}
\alias{filter_edges}
\title{Filter out edges which have weight more than threshold}
\usage{
filter_edges(edges, threshold)
}
\arguments{
\item{threshold}{}
}
\value{

}
\description{
Filter out edges which have weight more than threshold
}
