% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_markers.R
\name{filter_mk}
\alias{filter_mk}
\title{Filtering signals by background comparison information}
\usage{
filter_mk(
  gene_to_gene_cor,
  sig_summary,
  threshold = 0.1,
  pValue = 0.01,
  p_adjusted = TRUE,
  na_bg = TRUE
)
}
\arguments{
\item{p_adjusted}{use adjusted p-value or not. Default TRUE, use adjusted p-value.}

\item{na_bg}{default TRUE - change NA in the background and contrast correlation to TRUE or "pass" value.}
}
\value{

}
\description{
Filtering signals by background comparison information
}
