% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_panel_analysis.R
\name{get_abund_cross_panel}
\alias{get_abund_cross_panel}
\title{Compiling the matrix with cluster abundance data for all panels}
\usage{
get_abund_cross_panel(cell_clustering_list_set, use_samples)
}
\arguments{
\item{use_samples}{}
}
\value{

}
\description{
Compiling the matrix with cluster abundance data for all panels
}
