% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{get_abundance}
\alias{get_abundance}
\title{Get data of cluster abundace}
\usage{
get_abundance(list_cell_ctDist)
}
\arguments{
\item{list_cell_ctDist}{}
}
\value{

}
\description{
Get data of cluster abundace
}
