% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{get_abundance_correlation}
\alias{get_abundance_correlation}
\title{Calculation of correlation between cluster by abundance data}
\usage{
get_abundance_correlation(abundence_data, method = "spearman")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Calculation of correlation between cluster by abundance data
}
