% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusterisation.R
\name{get_abundance_dataframe}
\alias{get_abundance_dataframe}
\title{Create the data table to draw the abundance barplot}
\usage{
get_abundance_dataframe(fcs_raw, cell_clustering)
}
\arguments{
\item{cell_clustering}{}
}
\value{

}
\description{
Create the data table to draw the abundance barplot
}
