% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{get_adj_corr_pValue_matrix}
\alias{get_adj_corr_pValue_matrix}
\title{Extraction adjusted p-value from abundance correlation data}
\usage{
get_adj_corr_pValue_matrix(abundance_correlation)
}
\arguments{
\item{abundance_correlation}{}
}
\value{

}
\description{
Extraction adjusted p-value from abundance correlation data
}
