% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_markers.R
\name{get_anova_mk}
\alias{get_anova_mk}
\title{Get anova_mk object with ANOVA statistics for markers within each cluster across samples}
\usage{
get_anova_mk(list_cell_ctDist, list_expData, use_markers = NULL)
}
\arguments{
\item{use_markers}{}
}
\value{

}
\description{
Get anova_mk object with ANOVA statistics for markers within each cluster across samples
}
