% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_markers.R
\name{get_bg_anova_land_mk}
\alias{get_bg_anova_land_mk}
\title{Get bg_anova_land_mk object with background ANOVA statistics for markers across samples}
\usage{
get_bg_anova_land_mk(list_expData, use_markers = NULL)
}
\arguments{
\item{use_markers}{}
}
\value{

}
\description{
Get bg_anova_land_mk object with background ANOVA statistics for markers across samples
}
