% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_markers.R
\name{get_bg_corr_land_mk}
\alias{get_bg_corr_land_mk}
\title{Calculate background correlation landscape between markers across samples}
\usage{
get_bg_corr_land_mk(list_expData, use_markers = NULL, method = "spearman")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Calculate background correlation landscape between markers across samples
}
