% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{get_cell_annotation_fcs_files}
\alias{get_cell_annotation_fcs_files}
\title{Adding of cluster-info to flowSet object}
\usage{
get_cell_annotation_fcs_files(fcs_raw, cell_annotation, column_names = NULL)
}
\arguments{
\item{cell_clustering_list}{}
}
\value{

}
\description{
Adding of cluster-info to flowSet object
}
