% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusterisation.R
\name{get_cell_clustering_vector}
\alias{get_cell_clustering_vector}
\title{Create the cluster annotation as one vector}
\usage{
get_cell_clustering_vector(som, mc, k)
}
\arguments{
\item{k}{}
}
\value{

}
\description{
Create the cluster annotation as one vector
}
