% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{get_cell_type_from_gates}
\alias{get_cell_type_from_gates}
\title{Function adds set of gates to cell annotation object}
\usage{
get_cell_type_from_gates(gates, gate_list, cell_annotation, method = "pure")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Function adds set of gates to cell annotation object
}
