% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusterisation.R
\name{get_cluster_annotation}
\alias{get_cluster_annotation}
\title{Create the cluster annotation separated by samples}
\usage{
get_cluster_annotation(fcs_raw, som, mc, k)
}
\arguments{
\item{k}{}
}
\value{

}
\description{
Create the cluster annotation separated by samples
}
