% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_panel_analysis.R
\name{get_clustering_exp_cell_f_cross_panel}
\alias{get_clustering_exp_cell_f_cross_panel}
\title{Get dxpressing cell fraction data for target marker}
\usage{
get_clustering_exp_cell_f_cross_panel(
  fcs_raw_set,
  cell_clustering_list_set,
  use_samples,
  use_marker_set,
  target_mk,
  min_cell_number = 5
)
}
\arguments{
\item{min_cell_number}{}
}
\value{

}
\description{
Get dxpressing cell fraction data for target marker
}
