% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_panel_analysis.R
\name{get_common_markers}
\alias{get_common_markers}
\title{Extracting of common markers across panels and formed a comparison table}
\usage{
get_common_markers(use_marker_set)
}
\arguments{
\item{use_marker_set}{}
}
\value{

}
\description{
Extracting of common markers across panels and formed a comparison table
}
