% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusterisation.R
\name{get_consensusClust}
\alias{get_consensusClust}
\title{Get mc consensusCluster object from ConsensusClusterPlus}
\usage{
get_consensusClust(som, maxK = 20)
}
\arguments{
\item{maxK}{}
}
\value{

}
\description{
Get mc consensusCluster object from ConsensusClusterPlus
}
