% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_markers.R
\name{get_contrast_naive_corr_land_mk}
\alias{get_contrast_naive_corr_land_mk}
\title{Get contrast naive correlation landscape for each clone each marker versus each marker of the entire cell mass}
\usage{
get_contrast_naive_corr_land_mk(
  list_cell_ctDist,
  list_expData,
  use_markers = NULL,
  method = "spearman"
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Get contrast naive correlation landscape for each clone each marker versus each marker of the entire cell mass
}
