% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{get_data_for_gating}
\alias{get_data_for_gating}
\title{Take data to gating scatterplot}
\usage{
get_data_for_gating(fcs_raw, gating_subset, gating_mk1, gating_mk2)
}
\arguments{
\item{gating_mk2}{}
}
\value{

}
\description{
Take data to gating scatterplot
}
