% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_expression.R
\name{get_deconvol_dataframe}
\alias{get_deconvol_dataframe}
\title{Get dataframe for deconvolution plot of markers to samples-clusters projection}
\usage{
get_deconvol_dataframe(
  fcs_raw,
  use_markers,
  cell_clustering,
  summarize_mode = "median"
)
}
\arguments{
\item{cell_clustering}{}
}
\value{

}
\description{
Get dataframe for deconvolution plot of markers to samples-clusters projection
}
