% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_entire_panel}
\alias{get_entire_panel}
\title{Create entire named vector with obsermations: markers,clusters, qc etc.}
\usage{
get_entire_panel(fcs_raw)
}
\arguments{
\item{fcs_raw}{}
}
\value{

}
\description{
Create entire named vector with obsermations: markers,clusters, qc etc.
}
