% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusterisation.R
\name{get_fcs_cell_clustering_list}
\alias{get_fcs_cell_clustering_list}
\title{Create the cluster annotation as one vector for cluster info extraction from fcs file}
\usage{
get_fcs_cell_clustering_list(cell_clustering_list)
}
\arguments{
\item{cell_clustering_list}{}
}
\value{

}
\description{
Create the cluster annotation as one vector for cluster info extraction from fcs file
}
