% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_fcs_metadata}
\alias{get_fcs_metadata}
\title{Create basic metadata for flowSet object}
\usage{
get_fcs_metadata(fcs_files)
}
\arguments{
\item{fcs_files}{list with paths}
}
\value{
The dataframe with paths to fcs files within one column and names of these files in another column
}
\description{
The function creates two-column dataframe with paths
and names of files by the list of paths to fcs files with CyTOF data.
this data frame use as base metadata or phenoData in flowSet object
from flowCore package.
}
