% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_fcs_panel}
\alias{get_fcs_panel}
\title{Create panel data to fowSet object}
\usage{
get_fcs_panel(fcs_raw)
}
\arguments{
\item{fcs_raw}{flowSet object from flowCore package}
}
\value{
data frame with information about the markers
}
\description{
extract panel data about markers from flowSet
object from flowCore package. The function tries to assign
a description to markers and its type (technical or not)
}
