% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_fcs_raw}
\alias{get_fcs_raw}
\title{Creat FlowSet object}
\usage{
get_fcs_raw(md)
}
\arguments{
\item{md}{The input metadata should be dataframe format with
a column named path. The function assumed to process the result
of the function \dontrun{
get_fcs_metadata
}}
}
\value{
flowSet object from flowCore package
}
\description{
The function takes dataframe with column contained  paths to
fcs files and creates flowSet object from flowCore package.
}
