% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusterisation.R
\name{get_inds_subset}
\alias{get_inds_subset}
\title{Get indexs of rows with were sampled to subset}
\usage{
get_inds_subset(fcs_raw, sampling_size = 0.5, size_fuse = 5000)
}
\arguments{
\item{plot_ncell}{}
}
\value{

}
\description{
Get indexs of rows with were sampled to subset
}
