% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{get_list_cell_ctDist}
\alias{get_list_cell_ctDist}
\title{Get list_cell_ctDist object to correlation analysis}
\usage{
get_list_cell_ctDist(cell_clustering_list)
}
\arguments{
\item{cell_clustering_list}{}
}
\value{

}
\description{
Get list_cell_ctDist object to correlation analysis
}
