% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{get_list_expData}
\alias{get_list_expData}
\title{Get list_expData object to correlation analysis}
\usage{
get_list_expData(fcs_raw)
}
\arguments{
\item{fcs_raw}{}
}
\value{

}
\description{
Get list_expData object to correlation analysis
}
