% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_panel_analysis.R
\name{get_matrices_corr_info}
\alias{get_matrices_corr_info}
\title{Produce three matrices with corr coefficient, p-value and adjusted p-value for abundance correlations within cross panel assay}
\usage{
get_matrices_corr_info(
  abund_cross_p,
  method_cortest = "spearman",
  method_padj = "BH"
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Produce three matrices with corr coefficient, p-value and adjusted p-value for abundance correlations within cross panel assay
}
