% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_expression.R
\name{get_mk_clust_heatmap_dataframe}
\alias{get_mk_clust_heatmap_dataframe}
\title{Get dataframe for marker-cluster heatmap}
\usage{
get_mk_clust_heatmap_dataframe(
  fcs_raw,
  use_markers,
  cell_clustering,
  summarize_mode = "median"
)
}
\arguments{
\item{cell_clustering}{}
}
\value{

}
\description{
Get dataframe for marker-cluster heatmap
}
