% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDB.R
\name{get_neo_api}
\alias{get_neo_api}
\title{Create api for neo4j session}
\usage{
get_neo_api(user = "neo4j", password = "password")
}
\arguments{
\item{password}{}
}
\value{

}
\description{
Create api for neo4j session
}
