% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusterisation.R
\name{get_optimal_clusters}
\alias{get_optimal_clusters}
\title{Get a optimal number of clusters}
\usage{
get_optimal_clusters(mc, rate_var_expl = 0.9)
}
\arguments{
\item{rate_var_expl}{}
}
\value{

}
\description{
Get a optimal number of clusters
}
