% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_panel_analysis.R
\name{get_panel_content}
\alias{get_panel_content}
\title{Matrix for plotting the content of panels}
\usage{
get_panel_content(fcs_raw_set)
}
\arguments{
\item{md_set}{}
}
\value{

}
\description{
Matrix for plotting the content of panels
}
