% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_markers.R
\name{get_signal_in_cluster_mk}
\alias{get_signal_in_cluster_mk}
\title{Extract signals which go from and to the same cluster}
\usage{
get_signal_in_cluster_mk(signals)
}
\arguments{
\item{signals}{}
}
\value{

}
\description{
Extract signals which go from and to the same cluster
}
