% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{get_signals_between_clusters}
\alias{get_signals_between_clusters}
\title{Convert from signal dataframe to tidy dataframe with correlations between different clusters}
\usage{
get_signals_between_clusters(signals)
}
\arguments{
\item{signals}{}
}
\value{

}
\description{
Convert from signal dataframe to tidy dataframe with correlations between different clusters
}
