% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{get_signals_in_cluster}
\alias{get_signals_in_cluster}
\title{Convert from signal dataframe to tidy dataframe with correlations within one cluster}
\usage{
get_signals_in_cluster(signals)
}
\arguments{
\item{signals}{}
}
\value{

}
\description{
Convert from signal dataframe to tidy dataframe with correlations within one cluster
}
