% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{get_signals_in_cluster_top}
\alias{get_signals_in_cluster_top}
\title{Get the top of significant signals from signals_between_clusters object}
\usage{
get_signals_in_cluster_top(signals_in_cluster_top, use_markers)
}
\arguments{
\item{use_markers}{}
}
\value{

}
\description{
Get the top of significant signals from signals_between_clusters object
}
