% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_test_fcs_metadata}
\alias{get_test_fcs_metadata}
\title{extract metadata for build-in test dataset}
\usage{
get_test_fcs_metadata(test_data_dproc)
}
\arguments{
\item{test_data_dproc}{}
}
\value{

}
\description{
extract metadata for build-in test dataset
}
