% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_panel_analysis.R
\name{get_threshold_exp_cell_f_cross_panel}
\alias{get_threshold_exp_cell_f_cross_panel}
\title{Extract expression data of one marker for coss-panel analysis as one vector}
\usage{
get_threshold_exp_cell_f_cross_panel(
  fcs_raw_set,
  cell_clustering_list_set,
  use_samples,
  use_marker_set,
  target_mk,
  threshold = 20,
  min_cell_number = 5
)
}
\arguments{
\item{min_cell_number}{}
}
\value{

}
\description{
Extract expression data of one marker for coss-panel analysis as one vector
}
