% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_use_marker}
\alias{get_use_marker}
\title{filtering out technical markers}
\usage{
get_use_marker(panel)
}
\arguments{
\item{panel}{dataframe with marker panel data}
}
\value{
the list value of which are marker names and names of the
list element are names of antibody, relevant to the markers
}
\description{
the function formed list of markers which have class
mentioned as "type" in the panel. Other markers with class "state"
are considered as technical signals
}
