% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_panel_analysis.R
\name{get_write_corr_info_cross_p}
\alias{get_write_corr_info_cross_p}
\title{Preparing the table with cross-panel correlations for downloading}
\usage{
get_write_corr_info_cross_p(abund_corr_info)
}
\arguments{
\item{abund_corr_info}{}
}
\value{

}
\description{
Preparing the table with cross-panel correlations for downloading
}
