% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{htest_data_extractor}
\alias{htest_data_extractor}
\title{Low-level function for correlation analysis of genes from all clusters with abundance of one cluster}
\usage{
htest_data_extractor(abund_data, list_tt_expData, method = "spearman")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Low-level function for correlation analysis of genes from all clusters with abundance of one cluster
}
