% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_markers.R
\name{htest_data_extractor_mk}
\alias{htest_data_extractor_mk}
\title{Correlation analysis between markers of different clusters}
\usage{
htest_data_extractor_mk(
  list_tt_expData,
  use_markers = NULL,
  method = "spearman"
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Correlation analysis between markers of different clusters
}
