% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{sampled_tSNE}
\alias{sampled_tSNE}
\title{Preparing data to tSNE}
\usage{
sampled_tSNE(
  fcs_raw,
  use_markers,
  sampling_size = 0.5,
  method = "tSNE",
  perplexity = 30,
  theta = 0.5,
  max_iter = 1000
)
}
\arguments{
\item{sampling_size}{}
}
\value{

}
\description{
Preparing data to tSNE
}
