% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{signal_extractor}
\alias{signal_extractor}
\title{Get signal_Stat object with complete signal information from correlation analysis}
\usage{
signal_extractor(
  list_cell_ctDist,
  list_tt_expData,
  bg_ctCor_data,
  bg_anova,
  method = "spearman",
  bg_anova_alpha = 0.01,
  bg_ctCor_alpha = 0.01,
  gene_ctCor_alpha = 0.01
)
}
\arguments{
\item{gene_ctCor_alpha}{}
}
\value{

}
\description{
Get signal_Stat object with complete signal information from correlation analysis
}
