% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{signal_filter}
\alias{signal_filter}
\title{The first function for filtration of signals}
\usage{
signal_filter(signal_Stat, pValue = 0.01, threshold = 0.1)
}
\arguments{
\item{threshold}{}
}
\value{

}
\description{
The first function for filtration of signals
}
