% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{take_top_correlation}
\alias{take_top_correlation}
\title{Low-level function for extraction of top significant corrlation signals}
\usage{
take_top_correlation(cor_tdata)
}
\arguments{
\item{cor_tdata}{}
}
\value{

}
\description{
Low-level function for extraction of top significant corrlation signals
}
