% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{target_type_cortest}
\alias{target_type_cortest}
\title{Low-level function for correlation analysis of genes from one cluster with abundance of another cluster}
\usage{
target_type_cortest(
  target_type,
  abund_data,
  list_tt_expData,
  method = "spearman"
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Low-level function for correlation analysis of genes from one cluster with abundance of another cluster
}
