% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{target_type_data_extractor}
\alias{target_type_data_extractor}
\title{Low-level function to format detected signals to result dataframe}
\usage{
target_type_data_extractor(gene_ctDistr, bg_ctCor_data)
}
\arguments{
\item{bg_ctCor_data}{}
}
\value{

}
\description{
Low-level function to format detected signals to result dataframe
}
