% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{target_type_organiser}
\alias{target_type_organiser}
\title{Low-level function to get tt_expData object for one sample}
\usage{
target_type_organiser(cell_ctDist, expData)
}
\arguments{
\item{expData}{}
}
\value{

}
\description{
Low-level function to get tt_expData object for one sample
}
