% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{tt_sample_aggregator}
\alias{tt_sample_aggregator}
\title{Get list_tt_expData object for correlation analysis}
\usage{
tt_sample_aggregator(list_cell_ctDist, list_expData)
}
\arguments{
\item{list_expData}{}
}
\value{

}
\description{
Get list_tt_expData object for correlation analysis
}
