% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_old.R
\name{cytofCoreGUI}
\alias{cytofCoreGUI}
\title{The main function to run cutofCore graphic interface}
\usage{
cytofCoreGUI()
}
\value{
The function runs the Shiny App in browser.
}
\description{
The function creates two-column dataframe with paths
and names of files by the list of paths to fcs files with CyTOF data.
this data frame use as base metadata or phenoData in flowSet object
from flowCore package.
}
